+++ src/routes/api/article.js
@@ -165,7 +165,7 @@ router.route('/')
 	})
 	.post(function(req, res, next) {
 		res.json(req.results);
-		mail.sendThankYouMail(req.results, req.user, req.site) // todo: optional met config?
+		mail.sendThankYouMail(req.results, 'articles', req.user) // todo: optional met config?
 	})
 
 // one article